Minor fixes
-build - Added DOCKRPM_DEBUG and DOCKRPM_BASH environment variables.
Instead of rewriting the CMD command each time certain arguments are
changes (or docker is debugged), the environment variables are
overwritten
-check_dep - Fixed corner case where no dependencies resulted in a
grep return non-zero and failing dependency checking
-dockrpm/build - Removed dist from command line argument and added an
rpmmacros files. The rpmmacros files is NOT overridden every build.
This is the beginning of giving the user more control over
customization
-dockrpm - replaces message asserts with exit (should be exceptions)
-dockrpm - fake - missing newline
-dockrpm - code cleanup