Fixed deplist.py permissions
-docker_other - changed image name to be static for all packages. The
Only thing that can change it the UID/GID, and that's unlikely to change
If it does, it'll just rebuild each time. This WOULD be a race condition
if two users were building at the same time, but that wasn't solved before
either. Maybe the uid/gid should be added to the image name instead
-depcheck - Moved deplist lower on dockerfile, in case it is modified